Removed slapdtest module
