--- Doc/reference/slapdtest.rst
@@ -1,32 +1,0 @@
-###################################################################
-:mod:`slapdtest` Spawning test instances of OpenLDAP's slapd server
-###################################################################
-
-.. py:module:: slapdtest
-   :synopsis: Spawning test instances of OpenLDAP's slapd server
-.. moduleauthor:: python-ldap project (see https://www.python-ldap.org/)
-
-
-The module is used for testing python-ldap itself but can be used for
-automatically testing any OpenLDAP-based configuration setup.
-
-This module is pure Python and does not rely on any non-standard modules.
-Therefore it can be used stand-alone without the rest of the python-ldap
-package.
-
-Test fixtures for the popular `pytest` framework are developed in an external project, `pytest-ldap`_.
-
-Functions
-^^^^^^^^^
-
-
-Classes
-^^^^^^^
-
-.. autoclass:: slapdtest.SlapdObject
-   :members:
-
-.. autoclass:: slapdtest.SlapdTestCase
-   :members:
-
-.. _pytest-ldap: https://pypi.org/project/pytest-ldap/
