--- Lib/slapdtest/certs/ca.pem
@@ -1,80 +1,0 @@
-Certificate:
-    Data:
-        Version: 3 (0x2)
-        Serial Number: 1 (0x1)
-    Signature Algorithm: sha256WithRSAEncryption
-        Issuer: C=DE, O=python-ldap, OU=slapd-test, CN=Python LDAP Test CA
-        Validity
-            Not Before: Apr 12 18:52:38 2019 GMT
-            Not After : Oct 17 18:52:38 2994 GMT
-        Subject: C=DE, O=python-ldap, OU=slapd-test, CN=Python LDAP Test CA
-        Subject Public Key Info:
-            Public Key Algorithm: rsaEncryption
-                Public-Key: (2048 bit)
-                Modulus:
-                    00:d7:30:73:20:44:7d:83:d4:c7:01:b8:ab:1e:7c:
-                    91:f4:38:ac:9c:41:43:64:0c:31:99:48:70:22:7d:
-                    ae:1b:47:e7:2a:28:4d:f7:46:4e:b4:ba:ae:c0:9d:
-                    d5:1f:4b:7a:79:2f:b9:dc:68:7f:79:84:88:50:51:
-                    3b:7d:dc:d5:57:17:66:45:c0:2c:20:13:f7:99:d6:
-                    9d:e2:12:7c:41:76:82:51:19:2c:b6:ff:46:cb:04:
-                    56:38:22:2a:c3:7a:b5:71:51:49:4e:62:68:a0:99:
-                    6f:de:f3:a2:0f:a2:aa:1b:72:a5:87:bc:42:5a:a7:
-                    22:8d:33:b4:88:a8:dc:5d:72:ca:dd:a0:9a:4e:db:
-                    7d:8b:10:de:c5:41:e9:e9:8d:fa:6c:dd:94:6e:b1:
-                    31:c2:6d:a1:69:6c:7a:3a:b2:76:65:c9:e5:95:38:
-                    62:40:81:c6:29:26:26:d1:d1:c1:f4:5e:fa:24:ef:
-                    13:da:24:13:6f:f5:5c:ba:b1:31:8f:30:94:71:7b:
-                    c6:e5:da:b9:b5:64:39:39:09:c2:4a:80:64:58:1d:
-                    99:f5:65:3c:a7:26:08:95:26:35:7b:fa:e7:20:08:
-                    ff:72:df:9b:8f:9f:da:8b:c3:a7:8b:fc:8c:c0:a5:
-                    31:87:1d:4c:14:f6:cf:90:5e:2e:6e:a6:db:27:08:
-                    eb:df
-                Exponent: 65537 (0x10001)
-        X509v3 extensions:
-            X509v3 Basic Constraints: critical
-                CA:TRUE
-            X509v3 Key Usage: critical
-                Certificate Sign, CRL Sign
-            X509v3 Subject Key Identifier: 
-                BD:78:D5:4A:F1:90:96:C5:E8:EC:66:49:23:47:03:5F:26:73:86:B2
-            X509v3 Authority Key Identifier: 
-                keyid:BD:78:D5:4A:F1:90:96:C5:E8:EC:66:49:23:47:03:5F:26:73:86:B2
-
-    Signature Algorithm: sha256WithRSAEncryption
-         06:20:1f:eb:42:6a:42:62:b1:ee:69:c8:cd:47:a6:2e:69:95:
-         59:dc:49:09:69:40:93:25:a1:ec:6d:3a:dd:dc:e5:74:ab:33:
-         9d:8f:cc:e3:bb:7a:3f:5b:51:58:74:f7:bd:6c:7c:3c:b6:5a:
-         05:50:a8:8c:c3:fb:5b:75:2a:c2:6c:06:93:4c:a9:93:71:1c:
-         51:e5:be:a1:24:93:e2:79:ca:ea:08:86:90:b9:70:e7:7a:40:
-         bf:f4:d6:71:f4:4d:c0:0f:e0:31:a0:23:46:77:30:72:a9:62:
-         8a:2a:12:c4:dd:3d:86:ae:f7:6b:33:80:26:58:49:53:ff:cd:
-         8a:c6:f6:11:2c:b3:ff:a5:8e:1c:f8:22:e2:1b:8e:04:33:fb:
-         0d:da:31:86:12:9f:d1:03:86:9c:6a:78:5e:3c:5e:8a:52:aa:
-         68:1f:ff:f9:17:75:b0:da:f2:99:3c:80:3c:96:2a:33:07:54:
-         59:84:e7:92:34:0f:99:76:e3:d6:4d:4d:9c:fb:21:35:f9:cb:
-         a5:30:80:8b:9d:61:90:d3:d4:59:3a:2f:f2:f6:20:13:7e:26:
-         dc:50:b0:49:3e:19:fe:eb:7d:cf:b9:1a:5d:5c:3a:76:30:d9:
-         0e:d7:df:de:ce:a9:c4:21:df:63:b9:d0:64:86:0b:28:9a:2e:
-         ab:51:73:e4
------BEGIN CERTIFICATE-----
-MIIDjDCCAnSgAwIBAgIBATANBgkqhkiG9w0BAQsFADBWMQswCQYDVQQGEwJERTEU
-MBIGA1UECgwLcHl0aG9uLWxkYXAxEzARBgNVBAsMCnNsYXBkLXRlc3QxHDAaBgNV
-BAMME1B5dGhvbiBMREFQIFRlc3QgQ0EwIBcNMTkwNDEyMTg1MjM4WhgPMjk5NDEw
-MTcxODUyMzhaMFYxCzAJBgNVBAYTAkRFMRQwEgYDVQQKDAtweXRob24tbGRhcDET
-MBEGA1UECwwKc2xhcGQtdGVzdDEcMBoGA1UEAwwTUHl0aG9uIExEQVAgVGVzdCBD
-QTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBANcwcyBEfYPUxwG4qx58
-kfQ4rJxBQ2QMMZlIcCJ9rhtH5yooTfdGTrS6rsCd1R9Lenkvudxof3mEiFBRO33c
-1VcXZkXALCAT95nWneISfEF2glEZLLb/RssEVjgiKsN6tXFRSU5iaKCZb97zog+i
-qhtypYe8QlqnIo0ztIio3F1yyt2gmk7bfYsQ3sVB6emN+mzdlG6xMcJtoWlsejqy
-dmXJ5ZU4YkCBxikmJtHRwfRe+iTvE9okE2/1XLqxMY8wlHF7xuXaubVkOTkJwkqA
-ZFgdmfVlPKcmCJUmNXv65yAI/3Lfm4+f2ovDp4v8jMClMYcdTBT2z5BeLm6m2ycI
-698CAwEAAaNjMGEwDwYDVR0TAQH/BAUwAwEB/zAOBgNVHQ8BAf8EBAMCAQYwHQYD
-VR0OBBYEFL141UrxkJbF6OxmSSNHA18mc4ayMB8GA1UdIwQYMBaAFL141UrxkJbF
-6OxmSSNHA18mc4ayMA0GCSqGSIb3DQEBCwUAA4IBAQAGIB/rQmpCYrHuacjNR6Yu
-aZVZ3EkJaUCTJaHsbTrd3OV0qzOdj8zju3o/W1FYdPe9bHw8tloFUKiMw/tbdSrC
-bAaTTKmTcRxR5b6hJJPiecrqCIaQuXDnekC/9NZx9E3AD+AxoCNGdzByqWKKKhLE
-3T2GrvdrM4AmWElT/82KxvYRLLP/pY4c+CLiG44EM/sN2jGGEp/RA4acanhePF6K
-UqpoH//5F3Ww2vKZPIA8liozB1RZhOeSNA+ZduPWTU2c+yE1+culMICLnWGQ09RZ
-Oi/y9iATfibcULBJPhn+633PuRpdXDp2MNkO19/ezqnEId9judBkhgsomi6rUXPk
------END CERTIFICATE-----
