--- Lib/slapdtest/certs/client.pem
@@ -1,83 +1,0 @@
-Certificate:
-    Data:
-        Version: 3 (0x2)
-        Serial Number: 3 (0x3)
-    Signature Algorithm: sha256WithRSAEncryption
-        Issuer: C=DE, O=python-ldap, OU=slapd-test, CN=Python LDAP Test CA
-        Validity
-            Not Before: Apr 12 18:52:38 2019 GMT
-            Not After : Mar  1 18:52:38 3019 GMT
-        Subject: C=DE, O=python-ldap, OU=slapd-test, CN=client
-        Subject Public Key Info:
-            Public Key Algorithm: rsaEncryption
-                Public-Key: (2048 bit)
-                Modulus:
-                    00:e3:b7:93:ba:9d:1a:e7:01:63:e6:4f:42:6f:44:
-                    b0:47:31:c9:65:97:dd:88:58:5f:ce:e0:61:1d:70:
-                    9e:5f:6c:1c:4f:dc:1c:4a:d0:3b:fe:e9:a4:5c:47:
-                    b6:f6:a3:47:fa:7e:4f:db:80:d7:a9:5d:fb:ff:31:
-                    70:d0:70:59:44:49:b7:e6:50:d1:ac:84:59:99:af:
-                    4b:23:3e:b6:0f:0f:54:74:e9:36:6c:fb:fe:d1:b0:
-                    c5:e7:c7:57:3b:f6:61:46:11:0e:30:ab:30:ce:04:
-                    72:90:c5:e0:e1:34:99:3e:05:c1:34:40:00:6c:a8:
-                    5f:00:03:65:c2:db:f7:4a:d7:84:02:96:32:8b:0b:
-                    ce:24:d7:34:c5:4e:78:7c:76:e3:76:78:06:a5:68:
-                    f4:05:14:ea:23:e0:09:e8:9d:49:ec:c0:9e:39:cc:
-                    8c:79:57:60:5d:1b:ea:2a:3b:97:70:a3:f7:04:ba:
-                    42:a0:00:42:fc:4f:62:09:f4:2a:a5:9a:1d:ac:3d:
-                    b3:2a:60:2f:1c:38:3a:ba:61:cb:a0:33:39:f1:ad:
-                    f1:d5:e4:b3:53:85:f9:2f:c0:c7:b6:eb:eb:ef:eb:
-                    ca:fa:e7:36:a5:3c:3c:b4:33:e7:14:31:33:32:05:
-                    56:38:4c:bd:df:7a:3a:e5:a1:8d:ad:60:c8:4f:24:
-                    87:25
-                Exponent: 65537 (0x10001)
-        X509v3 extensions:
-            X509v3 Basic Constraints: critical
-                CA:FALSE
-            X509v3 Key Usage: critical
-                Digital Signature
-            X509v3 Extended Key Usage: critical
-                TLS Web Client Authentication
-            X509v3 Subject Key Identifier: 
-                4F:E7:35:C7:C8:C1:01:C3:7C:53:86:B9:BF:AE:8B:D6:45:A2:78:20
-            X509v3 Authority Key Identifier: 
-                keyid:BD:78:D5:4A:F1:90:96:C5:E8:EC:66:49:23:47:03:5F:26:73:86:B2
-
-    Signature Algorithm: sha256WithRSAEncryption
-         1c:90:5f:cf:18:48:95:4d:9d:d3:8e:6d:d1:69:19:1e:7b:3f:
-         1f:48:7c:c8:0d:2f:c4:53:0f:89:23:f4:be:ea:b4:7a:c6:dd:
-         cc:18:0f:e7:34:ea:2c:d4:07:0d:65:78:e8:20:40:3f:36:ef:
-         2c:00:31:69:e6:20:48:65:be:57:03:0e:69:ff:b9:83:59:99:
-         7d:4d:86:98:14:5b:8e:39:25:3a:a8:6d:51:dc:45:a5:0f:cd:
-         f3:7a:fd:55:af:5f:55:75:20:03:f5:4a:75:6a:79:2f:76:84:
-         f6:4e:3d:1d:59:45:9a:b1:6a:57:6f:16:76:76:f8:df:6e:96:
-         d5:25:27:34:4b:21:d8:c9:9a:36:55:45:a0:43:16:43:68:93:
-         37:af:81:89:06:d1:56:1b:9e:0f:62:40:ad:3c:4c:f5:ef:6c:
-         a2:a4:7f:f2:fa:78:9c:0d:c0:19:f1:10:e8:d8:cf:03:67:3c:
-         2d:4d:f3:5d:67:5c:41:a7:4f:d6:c5:0e:ff:2c:04:dd:23:bb:
-         85:44:8e:25:ac:15:a3:82:fa:a4:4f:fa:1d:87:f0:58:dc:ae:
-         53:05:b9:81:e8:cb:e5:0c:ac:a5:74:68:03:f9:22:a0:45:b6:
-         62:58:e0:98:d9:8c:54:a4:22:03:7a:37:12:eb:7d:b1:ad:45:
-         60:8e:7a:df
------BEGIN CERTIFICATE-----
-MIIDlDCCAnygAwIBAgIBAzANBgkqhkiG9w0BAQsFADBWMQswCQYDVQQGEwJERTEU
-MBIGA1UECgwLcHl0aG9uLWxkYXAxEzARBgNVBAsMCnNsYXBkLXRlc3QxHDAaBgNV
-BAMME1B5dGhvbiBMREFQIFRlc3QgQ0EwIBcNMTkwNDEyMTg1MjM4WhgPMzAxOTAz
-MDExODUyMzhaMEkxCzAJBgNVBAYTAkRFMRQwEgYDVQQKDAtweXRob24tbGRhcDET
-MBEGA1UECwwKc2xhcGQtdGVzdDEPMA0GA1UEAwwGY2xpZW50MIIBIjANBgkqhkiG
-9w0BAQEFAAOCAQ8AMIIBCgKCAQEA47eTup0a5wFj5k9Cb0SwRzHJZZfdiFhfzuBh
-HXCeX2wcT9wcStA7/umkXEe29qNH+n5P24DXqV37/zFw0HBZREm35lDRrIRZma9L
-Iz62Dw9UdOk2bPv+0bDF58dXO/ZhRhEOMKswzgRykMXg4TSZPgXBNEAAbKhfAANl
-wtv3SteEApYyiwvOJNc0xU54fHbjdngGpWj0BRTqI+AJ6J1J7MCeOcyMeVdgXRvq
-KjuXcKP3BLpCoABC/E9iCfQqpZodrD2zKmAvHDg6umHLoDM58a3x1eSzU4X5L8DH
-tuvr7+vK+uc2pTw8tDPnFDEzMgVWOEy933o65aGNrWDITySHJQIDAQABo3gwdjAM
-BgNVHRMBAf8EAjAAMA4GA1UdDwEB/wQEAwIHgDAWBgNVHSUBAf8EDDAKBggrBgEF
-BQcDAjAdBgNVHQ4EFgQUT+c1x8jBAcN8U4a5v66L1kWieCAwHwYDVR0jBBgwFoAU
-vXjVSvGQlsXo7GZJI0cDXyZzhrIwDQYJKoZIhvcNAQELBQADggEBAByQX88YSJVN
-ndOObdFpGR57Px9IfMgNL8RTD4kj9L7qtHrG3cwYD+c06izUBw1leOggQD827ywA
-MWnmIEhlvlcDDmn/uYNZmX1NhpgUW445JTqobVHcRaUPzfN6/VWvX1V1IAP1SnVq
-eS92hPZOPR1ZRZqxaldvFnZ2+N9ultUlJzRLIdjJmjZVRaBDFkNokzevgYkG0VYb
-ng9iQK08TPXvbKKkf/L6eJwNwBnxEOjYzwNnPC1N811nXEGnT9bFDv8sBN0ju4VE
-jiWsFaOC+qRP+h2H8FjcrlMFuYHoy+UMrKV0aAP5IqBFtmJY4JjZjFSkIgN6NxLr
-fbGtRWCOet8=
------END CERTIFICATE-----
