--- Lib/slapdtest/certs/server.conf
@@ -1,16 +1,0 @@
-# Written by Christian Heimes
-
-[req]
-default_bits = 2048
-encrypt_key = no
-default_md = sha256
-utf8 = yes
-string_mask = utf8only
-prompt = no
-distinguished_name = server_dn
-
-[server_dn]
-countryName = "DE"
-organizationName = "python-ldap"
-organizationalUnitName = "slapd-test"
-commonName = "server cert for localhost"
