--- Lib/slapdtest/certs/server.pem
@@ -1,86 +1,0 @@
-Certificate:
-    Data:
-        Version: 3 (0x2)
-        Serial Number: 2 (0x2)
-    Signature Algorithm: sha256WithRSAEncryption
-        Issuer: C=DE, O=python-ldap, OU=slapd-test, CN=Python LDAP Test CA
-        Validity
-            Not Before: Apr 12 18:52:38 2019 GMT
-            Not After : Mar  1 18:52:38 3019 GMT
-        Subject: C=DE, O=python-ldap, OU=slapd-test, CN=server cert for localhost
-        Subject Public Key Info:
-            Public Key Algorithm: rsaEncryption
-                Public-Key: (2048 bit)
-                Modulus:
-                    00:ac:06:4d:26:97:71:11:14:9c:a0:23:ab:a3:20:
-                    96:04:45:55:38:76:84:d9:65:de:2b:e3:3f:cb:39:
-                    e7:f4:87:28:98:88:84:00:73:dc:1e:b5:aa:19:a4:
-                    cf:9b:97:81:21:7a:bc:72:f5:20:5a:08:1c:11:cd:
-                    e7:db:69:84:33:d9:fd:a1:74:40:db:fa:c2:12:3a:
-                    24:e4:7b:57:a2:17:6b:a0:60:d9:52:e7:95:a4:d4:
-                    a9:c2:9c:e4:44:f5:b8:dc:48:d2:cf:b2:c7:28:51:
-                    2c:c4:d9:b9:c1:4a:85:30:76:25:4a:d7:c6:0e:09:
-                    94:7a:19:37:82:f1:e0:37:7a:4b:e5:47:ef:60:f8:
-                    50:4e:81:0f:7e:4a:d5:cf:0d:da:be:88:80:06:5d:
-                    94:44:24:8b:9a:9a:e9:c4:ae:d4:ce:c3:ce:43:7b:
-                    86:05:0a:7c:c6:c3:75:1e:08:84:c3:0f:8c:05:98:
-                    8d:88:5a:98:51:17:05:8f:b1:20:f0:50:97:c0:4c:
-                    0f:12:fb:b4:66:e2:fc:2b:42:3e:f6:ea:05:03:96:
-                    27:13:cc:f5:ca:84:98:d7:8e:3e:97:2f:bc:57:1d:
-                    fa:50:c0:c3:e4:04:2a:f1:7a:c2:9c:78:73:09:fa:
-                    6a:d1:b2:eb:aa:1d:c8:ae:8b:14:62:83:de:26:20:
-                    b1:25
-                Exponent: 65537 (0x10001)
-        X509v3 extensions:
-            X509v3 Basic Constraints: critical
-                CA:FALSE
-            X509v3 Key Usage: critical
-                Digital Signature, Key Encipherment
-            X509v3 Extended Key Usage: critical
-                TLS Web Server Authentication
-            X509v3 Subject Key Identifier: 
-                08:D1:86:1B:82:0A:4F:71:31:E4:F5:31:23:CC:67:3B:FA:84:3B:A0
-            X509v3 Authority Key Identifier: 
-                keyid:BD:78:D5:4A:F1:90:96:C5:E8:EC:66:49:23:47:03:5F:26:73:86:B2
-
-            X509v3 Subject Alternative Name: 
-                DNS:localhost, IP Address:127.0.0.1, IP Address:0:0:0:0:0:0:0:1
-    Signature Algorithm: sha256WithRSAEncryption
-         88:60:af:be:11:c4:aa:dc:9b:f1:e7:14:da:20:aa:6f:2f:06:
-         ae:38:b2:7c:ac:90:81:22:51:7e:cb:26:15:6e:fe:67:98:c1:
-         0d:dc:aa:39:98:2b:d2:cc:3c:ff:1a:92:2f:56:0a:a9:6e:d8:
-         9a:3d:c5:4d:6f:cc:91:2e:e3:4e:bf:22:ab:cb:92:1a:a0:8f:
-         43:cd:82:bc:48:55:c4:95:cf:10:6b:6a:31:19:92:7d:e0:06:
-         05:6f:0b:33:e7:2a:37:42:f9:ec:1b:29:99:e1:58:0c:01:a7:
-         c3:8b:58:71:21:9f:61:8c:a7:fb:b6:7e:32:8b:a9:4e:c7:1f:
-         f6:46:e8:dd:ac:a6:4c:53:f8:4d:93:e4:ec:73:ab:0b:be:98:
-         c5:78:c4:92:c0:4c:78:47:52:2f:93:07:67:20:a4:5a:7f:59:
-         7e:4f:48:53:20:0d:37:bb:06:f8:44:42:64:b4:94:15:43:d1:
-         4c:51:f3:97:1d:2d:cd:db:b9:bb:1a:69:10:89:7d:ae:1d:0d:
-         94:78:45:29:cd:c4:42:67:67:96:05:bf:da:aa:23:65:7b:04:
-         ff:b7:ac:9d:ee:0b:e7:0f:c1:c5:0b:48:fe:0f:d6:3f:d8:b4:
-         77:12:bb:f5:91:4f:43:e6:01:3f:a4:c0:ea:8c:c6:68:99:8e:
-         49:e8:c4:8b
------BEGIN CERTIFICATE-----
-MIID1zCCAr+gAwIBAgIBAjANBgkqhkiG9w0BAQsFADBWMQswCQYDVQQGEwJERTEU
-MBIGA1UECgwLcHl0aG9uLWxkYXAxEzARBgNVBAsMCnNsYXBkLXRlc3QxHDAaBgNV
-BAMME1B5dGhvbiBMREFQIFRlc3QgQ0EwIBcNMTkwNDEyMTg1MjM4WhgPMzAxOTAz
-MDExODUyMzhaMFwxCzAJBgNVBAYTAkRFMRQwEgYDVQQKDAtweXRob24tbGRhcDET
-MBEGA1UECwwKc2xhcGQtdGVzdDEiMCAGA1UEAwwZc2VydmVyIGNlcnQgZm9yIGxv
-Y2FsaG9zdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAKwGTSaXcREU
-nKAjq6MglgRFVTh2hNll3ivjP8s55/SHKJiIhABz3B61qhmkz5uXgSF6vHL1IFoI
-HBHN59tphDPZ/aF0QNv6whI6JOR7V6IXa6Bg2VLnlaTUqcKc5ET1uNxI0s+yxyhR
-LMTZucFKhTB2JUrXxg4JlHoZN4Lx4Dd6S+VH72D4UE6BD35K1c8N2r6IgAZdlEQk
-i5qa6cSu1M7DzkN7hgUKfMbDdR4IhMMPjAWYjYhamFEXBY+xIPBQl8BMDxL7tGbi
-/CtCPvbqBQOWJxPM9cqEmNeOPpcvvFcd+lDAw+QEKvF6wpx4cwn6atGy66odyK6L
-FGKD3iYgsSUCAwEAAaOBpzCBpDAMBgNVHRMBAf8EAjAAMA4GA1UdDwEB/wQEAwIF
-oDAWBgNVHSUBAf8EDDAKBggrBgEFBQcDATAdBgNVHQ4EFgQUCNGGG4IKT3Ex5PUx
-I8xnO/qEO6AwHwYDVR0jBBgwFoAUvXjVSvGQlsXo7GZJI0cDXyZzhrIwLAYDVR0R
-BCUwI4IJbG9jYWxob3N0hwR/AAABhxAAAAAAAAAAAAAAAAAAAAABMA0GCSqGSIb3
-DQEBCwUAA4IBAQCIYK++EcSq3Jvx5xTaIKpvLwauOLJ8rJCBIlF+yyYVbv5nmMEN
-3Ko5mCvSzDz/GpIvVgqpbtiaPcVNb8yRLuNOvyKry5IaoI9DzYK8SFXElc8Qa2ox
-GZJ94AYFbwsz5yo3QvnsGymZ4VgMAafDi1hxIZ9hjKf7tn4yi6lOxx/2RujdrKZM
-U/hNk+Tsc6sLvpjFeMSSwEx4R1IvkwdnIKRaf1l+T0hTIA03uwb4REJktJQVQ9FM
-UfOXHS3N27m7GmkQiX2uHQ2UeEUpzcRCZ2eWBb/aqiNlewT/t6yd7gvnD8HFC0j+
-D9Y/2LR3Erv1kU9D5gE/pMDqjMZomY5J6MSL
------END CERTIFICATE-----
