+++ MANIFEST.in
@@ -5,6 +5,5 @@ recursive-include Build *.cfg*
 recursive-include Lib *.py
 recursive-include Demo *.py
 recursive-include Tests *.py *.ldif
-recursive-include Lib/slapdtest *.pem *.key *.conf *.sh README
 recursive-include Doc *.rst *.py spelling_wordlist.txt Makefile
 prune Doc/.build
