+++ setup.py
@@ -153,7 +153,6 @@ setup(
     'ldap.controls',
     'ldap.extop',
     'ldap.schema',
-    'slapdtest',
   ],
   package_dir = {'': 'Lib',},
   data_files = LDAP_CLASS.extra_files,
